+++ src/relstorage/adapters/postgresql/drivers/pg8000.py
@@ -115,17 +115,35 @@ class PG8000Driver(AbstractPostgreSQLDriver):
                              user, host='localhost',
                              unix_sock=None,
                              port=5432, database=None,
-                             password=None, ssl=None,
+                             password=None, ssl_context=None,
                              timeout=None, application_name=None,
                              max_prepared_statements=1000,
                              tcp_keepalive=True):
                     # pylint:disable=useless-super-delegation
-                    # We have to do this because the super class requires
-                    # all these arguments and doesn't have defaults
-                    super(Connection, self).__init__(
-                        user, host, unix_sock,
-                        port, database, password, ssl, timeout, application_name,
-                        max_prepared_statements, tcp_keepalive)
+
+                    # Prior to pg8000 1.15.1, we had to do this
+                    # because the super class requires all these
+                    # arguments and didn't have defaults. In that release, it gained
+                    # defaults, but our order had been wrong for some time. In
+                    # 1.15.3 our bad ordering started causing TypeError.
+                    # Pass these things in this way to be compatible with older versions too
+                    # that didn't have defaults or extra parameters, but the names
+                    # were the same...except for `ssl` vs `ssl_context`: the latter
+                    # is only in versions that run on Python 3.
+                    # We could probably do better than this
+                    kwargs = dict(
+                        user=user, host=host, unix_sock=unix_sock,
+                        port=port, database=database,
+                        password=password, ssl_context=ssl_context,
+                        timeout=timeout,
+                        application_name=application_name,
+                        max_prepared_statements=max_prepared_statements,
+                        tcp_keepalive=tcp_keepalive
+                    )
+                    if str is bytes: # PY2
+                        del kwargs['ssl_context']
+                        kwargs['ssl'] = ssl_context
+                    super(Connection, self).__init__(**kwargs)
 
                 def cursor(self):
                     return Cursor(self)